fix STREAM-UNREAD-CHAR
In STREAM-READ-CHAR, the first value returned from the
BABEL:CODE-POINT-COUNTER function was used to track
LAST-CHAR-SIZE (the number of bytes consumed while reading a
character). However, it is actually the number of characters decoded,
and for this particular call, it is always 1. As a result, a later
STREAM-UNREAD-CHAR of a multibyte character will cause BUFFER-POSITION
to point into the middle of the encoded byte sequence, ultimately
leading to a decoding error.

Instead, use the second value returned from the
BABEL:CODE-POINT-COUNTER function, the position in the buffer to start
the next decoding, and compute LAST-CHAR-SIZE using its difference
from BUFFER-POSITION.