Allow overriding content-type header when sending requests
For sending e.g. JSON payloads with POST requests, one needs to set
the proper content type (in this case, "application/json"). Currently,
this is not possible because this header is filtered out and dexador
tries to set it automatically, which does not work in these scenarios.

This patch allows to override the content type manually. If not present,
dexador will still try to guess it automatically.