Fix failure in auto-retry when connection-cache is used and server drops connection
Unsafe with-retrying macrolet uses variables from outside of them
(in particular: 'stream'), we were caught by 'stream' being locally
bound around one of the with-retrying clauses.  This caused the
auto-retry when re-using a connection pool connection when the
connection has been closed by the server to not work depending
on when the stream error occurred (I think the stream error usually
occurs before this clause, when writing the headers, but I had a
reproducible case where it did not).  The symptom was getting an
error that we were trying to reuse a closed SSL-STREAM.